+++ workers/tests/unittests/test_api_request_models.py
@@ -4,9 +4,15 @@ import json
 import os
 import unittest
 
-from geojson import Polygon
+import pytest
+from pydantic import ValidationError
 
 from ohsome_quality_analyst.api import request_models
+from ohsome_quality_analyst.utils.exceptions import (
+    GeoJSONError,
+    GeoJSONGeometryTypeError,
+    GeoJSONObjectTypeError,
+)
 
 
 class TestApiRequestModels(unittest.TestCase):
@@ -18,54 +24,6 @@ class TestApiRequestModels(unittest.TestCase):
         )
         with open(path, "r") as file:
             self.bpolys = json.load(file)
-
-    def test_base_indicator_valid(self):
-        request_models.BaseIndicator(name="minimal")
-        request_models.BaseIndicator(
-            name="minimal",
-            include_svg=True,
-            include_html=True,
-            include_data=False,
-            flatten=False,
-        )
-
-    def test_base_indicator_invalid(self):
-        with self.assertRaises(ValueError):
-            request_models.BaseIndicator()
-        with self.assertRaises(ValueError):
-            request_models.BaseIndicator(name="foo")
-        with self.assertRaises(ValueError):
-            request_models.BaseIndicator(includeSvg=True)
-        with self.assertRaises(ValueError):
-            request_models.BaseIndicator(name="minimal", include_svg="foo")
-        with self.assertRaises(ValueError):
-            request_models.BaseIndicator(name="minimal", include_html="foo")
-        with self.assertRaises(ValueError):
-            request_models.BaseIndicator(name="minimal", flatten="foo")
-
-    def test_base_report_valid(self):
-        request_models.BaseReport(name="minimal")
-        request_models.BaseReport(
-            name="minimal",
-            include_svg=True,
-            include_html=True,
-            include_data=False,
-            flatten=False,
-        )
-
-    def test_base_report_invalid(self):
-        with self.assertRaises(ValueError):
-            request_models.BaseReport()
-        with self.assertRaises(ValueError):
-            request_models.BaseReport(name="foo")
-        with self.assertRaises(ValueError):
-            request_models.BaseReport(include_svg=True)
-        with self.assertRaises(ValueError):
-            request_models.BaseReport(name="minimal", include_svg="foo")
-        with self.assertRaises(ValueError):
-            request_models.BaseReport(name="minimal", include_html="foo")
-        with self.assertRaises(ValueError):
-            request_models.BaseReport(name="minimal", flatten="foo")
 
     def test_topic_key_valid(self):
         # Test on BaseIndicator because validation of BaseTopic needs indicator name
@@ -111,69 +69,8 @@ class TestApiRequestModels(unittest.TestCase):
         with self.assertRaises(ValueError):
             request_models.BaseDatabase(dataset="foo", feature_id="3")
 
-    def test_bpolys_valid(self):
-        request_models.BaseBpolys(bpolys=self.bpolys)
-
-    def test_bpolys_invalid(self):
-        bpolys = Polygon(
-            [[(2.38, 57.322), (23.194, -20.28), (-120.43, 19.15), (2.0, 1.0)]]
-        )
-
-        with self.assertRaises(ValueError):
-            request_models.BaseBpolys(bpolys=bpolys)
-
-    def test_indicator_database(self):
-        request_models.IndicatorDatabase(
-            name="minimal",
-            topic="minimal",
-            dataset="regions",
-            feature_id="3",
-        )
-        request_models.IndicatorDatabase(
-            name="minimal",
-            topic="minimal",
-            dataset="regions",
-            feature_id="Heidelberg",
-            fid_field="name",
-        )
-
-    def test_indicator_bpolys(self):
-        request_models.IndicatorBpolys(
-            name="minimal",
-            topic="minimal",
-            bpolys=self.bpolys,
-        )
-
-    def test_indicator_invalid_topic_combination(self):
-        kwargs = {
-            "name": "minimal",
-            "topic": "amenities",
-            "dataset": "regions",
-            "feature-id": 3,
-        }
-        with self.assertRaises(ValueError):
-            request_models.IndicatorDatabase(**kwargs)
-        with self.assertRaises(ValueError):
-            request_models.IndicatorBpolys(**kwargs)
-
-    def test_indicator_data(self):
-        request_models.IndicatorData(
-            name="mapping-saturation",
-            bpolys=self.bpolys,
-            topic={"key": "foo", "name": "bar", "description": "buz", "data": {}},
-        )
-
-    def test_indicator_data_invalid_indicator(self):
-        with self.assertRaises(ValueError):
-            request_models.IndicatorData(
-                name="foo",
-                bpolys=self.bpolys,
-                topic={"key": "foo", "name": "bar", "description": "buz", "data": {}},
-            )
-
     def test_invalid_set_of_arguments(self):
         param_keys = (
-            "name",
             "topic",
             "dataset",
             "feature-id",
@@ -181,7 +78,6 @@ class TestApiRequestModels(unittest.TestCase):
             "bpolys",
         )
         param_values = (
-            "minimal",
             "minimal",
             "regions",
             "3",
@@ -199,20 +95,17 @@ class TestApiRequestModels(unittest.TestCase):
                 )
         valid_combinations = (
             {
-                "name": "minimal",
                 "topic": "minimal",
                 "dataset": "regions",
                 "feature-id": "3",
             },
             {
-                "name": "minimal",
                 "topic": "minimal",
                 "dataset": "regions",
                 "feature-id": "3",
                 "fid-field": "ogc_fid",
             },
             {
-                "name": "minimal",
                 "topic": "minimal",
                 "bpolys": self.bpolys,
             },
@@ -226,3 +119,107 @@ class TestApiRequestModels(unittest.TestCase):
             ):
                 with self.assertRaises(ValueError):
                     model(**combination)
+
+
+def test_base_indicator_valid():
+    request_models.BaseIndicator()
+    request_models.BaseIndicator(
+        include_svg=True,
+        include_html=True,
+        include_data=True,
+        flatten=True,
+    )
+
+
+def test_base_indicator_invalid():
+    with pytest.raises(ValueError):
+        request_models.BaseIndicator(include_svg="foo")
+    with pytest.raises(ValueError):
+        request_models.BaseIndicator(include_html="foo")
+    with pytest.raises(ValueError):
+        request_models.BaseIndicator(include_data="foo")
+    with pytest.raises(ValueError):
+        request_models.BaseIndicator(flatten="foo")
+
+
+def test_base_report_valid():
+    request_models.BaseReport()
+    request_models.BaseReport(
+        include_svg=True,
+        include_html=True,
+        include_data=False,
+        flatten=False,
+    )
+
+
+def test_base_report_invalid():
+    with pytest.raises(ValueError):
+        request_models.BaseReport(include_svg="foo")
+    with pytest.raises(ValueError):
+        request_models.BaseReport(include_html="foo")
+    with pytest.raises(ValueError):
+        request_models.BaseReport(include_data="foo")
+    with pytest.raises(ValueError):
+        request_models.BaseReport(flatten="foo")
+
+
+def test_bpolys_valid(
+    feature_collection_germany_heidelberg,
+    feature_collection_germany_heidelberg_bahnstadt_bergheim,
+):
+    # Single Feature
+    request_models.BaseBpolys(bpolys=feature_collection_germany_heidelberg)
+    # Multiple Features
+    request_models.BaseBpolys(
+        bpolys=feature_collection_germany_heidelberg_bahnstadt_bergheim
+    )
+
+
+def test_bpolys_invalid(feature_collection_invalid):
+    with pytest.raises((GeoJSONError, ValidationError)):
+        request_models.BaseBpolys(bpolys=feature_collection_invalid)
+
+
+# TODO
+@pytest.mark.skip(reason="Support for Feature will be discontinued.")
+def test_bpolys_unsupported_object_type_feature(feature_germany_heidelberg):
+    with pytest.raises((GeoJSONObjectTypeError, ValidationError)):
+        request_models.BaseBpolys(bpolys=feature_germany_heidelberg)
+
+
+def test_bpolys_unsupported_object_type(geojson_unsupported_object_type):
+    with pytest.raises((GeoJSONObjectTypeError, ValidationError)):
+        request_models.BaseBpolys(bpolys=geojson_unsupported_object_type)
+
+
+def test_bpolys_unsupported_geometry_type(feature_collection_unsupported_geometry_type):
+    with pytest.raises((GeoJSONGeometryTypeError, ValidationError)):
+        request_models.BaseBpolys(bpolys=feature_collection_unsupported_geometry_type)
+
+
+def test_indicator_database():
+    request_models.IndicatorDatabase(
+        topic="minimal",
+        dataset="regions",
+        feature_id="3",
+    )
+    request_models.IndicatorDatabase(
+        topic="minimal",
+        dataset="regions",
+        feature_id="Heidelberg",
+        fid_field="name",
+    )
+
+
+def test_indicator_bpolys(feature_collection_germany_heidelberg):
+    request_models.IndicatorBpolys(
+        topic="minimal",
+        bpolys=feature_collection_germany_heidelberg,
+    )
+
+
+def test_indicator_data(feature_collection_germany_heidelberg):
+    request_models.IndicatorData(
+        bpolys=feature_collection_germany_heidelberg,
+        topic={"key": "foo", "name": "bar", "description": "buz", "data": {}},
+    )