+++ workers/tests/unittests/test_report_minimal.py
@@ -3,13 +3,13 @@ from unittest.mock import Mock
 
 from ohsome_quality_analyst.reports.minimal.report import Minimal
 
-from .utils import get_geojson_fixture
+from ..utils import load_geojson_fixture
 
 
 class TestReportMinimal(unittest.TestCase):
     def test_combine_indicators_mean(self):
-        geometry = get_geojson_fixture("heidelberg-altstadt-geometry.geojson")
-        report = Minimal(geometry)
+        feature = load_geojson_fixture("feature-germany-heidelberg.geojson")
+        report = Minimal(feature)
 
         # Mock indicator objects with a fixed result value
         for _ in report.indicator_topic: